+++ src/main/kotlin/com/github/jairrab/calc/lib/mathutils/EntriesCalculator.kt
@@ -2,6 +2,8 @@
 
 package com.github.jairrab.calc.lib.mathutils
 
+import java.math.BigDecimal
+
 interface EntriesCalculator {
-    fun solve(entries: List<String>): Double
+    fun solve(entries: List<String>): BigDecimal
 }
